% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2.R
\docType{data}
\name{m2}
\alias{m2}
\title{CEATTLE estimates of natural mortality due to cannibalism}
\format{
\subsection{\code{m2}}{

A data frame with 1,815 rows and 4 columns:
\describe{
\item{name}{The name of the data, corresponding to the original csv name}
\item{age}{An integer describing the fish age}
\item{year}{A four-digit integer giving the year}
\item{m2}{A real number of cannibalistic natural mortality}
}
}
}
\source{
Sophia Wassermann
}
\usage{
m2
}
\description{
Estimates from the CEATTLE model of cannibalistic natural mortality for
ages 1--5 starting in 1980 from three different model runs. The three
model runs are provided because all diet data was averaged and provided
as a single data set of diets to CEATTLE, where early years had higher
instances of cannibalism and more recent years had much lower instances
of cannibalism. Thus the average over the entire time series is the
base model and the other variants represent the average over a specific
time frame rather than the entire data set.
}
\keyword{datasets}
